refact: alterações no src/core/types para ficar em conformidade com o bakcend
