+++ src/main/frontend/src/core/types.ts
@@ -16,9 +16,12 @@ export interface UsuarioRespostaDTO {
   nome: string;
   email: string;
   tpUsuario: TpUsuario;
-  situacao: Situacao;
   dtCadastro?: string;
   dtModificacao?: string;
+  situacao: Situacao;
+  agendamentoList?: { id: number }[];
+  eventoList?: { id: number }[];
+  pacienteList?: { id: number }[];
 }
 
 export interface SessionUser {
@@ -32,39 +35,45 @@ export interface SessionUser {
 export interface PacienteRespostaDTO {
   id: number;
   nome: string;
-  sobrenome?: string;
-  cpf?: string;
+  sobrenome: string;
+  cpf: string;
   dtNascimento: string;
   crtSus?: string;
   email?: string;
   contato?: string;
-  situacao: Situacao;
+  usuario?: { id: number };
   dtCadastro?: string;
   dtModificacao?: string;
-  usuario?: { id: number };
+  situacao: Situacao;
+  agendamentoList?: { id: number }[];
 }
 
 export interface PacientePayload {
   id?: number;
   nome: string;
-  sobrenome?: string;
-  cpf?: string;
+  sobrenome: string;
+  cpf: string;
   dtNascimento: string;
   crtSus?: string;
   email?: string;
   contato?: string;
-  situacao?: Situacao;
+  situacao: Situacao;
   idUsuario: number;
 }
-
 export interface AgendamentoRespostaDTO {
   id: number;
-  paciente: { id: number; nome: string; sobrenome?: string };
+  paciente: {
+    id: number;
+    nome: string;
+    sobrenome: string;
+  };
   tpVisita: TpVisita;
   dtAgendamento: string;
-  observacao?: string;
-  situacao: Situacao;
   usuario?: { id: number };
+  dtCadastro?: string;
+  dtModificacao?: string;
+  situacao: Situacao;
+  observacao?: string;
 }
 
 export interface AgendamentoPayload {
@@ -78,9 +87,11 @@ export interface AgendamentoPayload {
 
 export interface EventoRespostaDTO {
   id: number;
-  nmEvento?: string;
-  descricao?: string;
+  nmEvento: string;
+  descricao: string;
   dtEvento: string;
+  dtCadastro?: string;
+  dtModificacao?: string;
   situacao: Situacao;
   usuario?: { id: number };
 }
@@ -90,7 +101,7 @@ export interface EventoPayload {
   nmEvento: string;
   dtEvento: string;
   descricao?: string;
-  situacao?: Situacao;
+  //situação não é necessária no payload, pois o backend criação e atualização não possui 
   idUsuario: number;
 }
 
@@ -98,9 +109,8 @@ export interface UsuarioPayload {
   id?: number;
   nome: string;
   email: string;
-  senha?: string;
+  senha: string;
   tpUsuario: TpUsuario;
-  situacao?: Situacao;
 }
 
 export interface ArquivoMeta {
